+++ src/VBox/Frontends/VBoxManage/VBoxManageGuestCtrl.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxManageGuestCtrl.cpp 107828 2025-01-14 13:10:55Z andreas.loeffler@oracle.com $ */
+/* $Id: VBoxManageGuestCtrl.cpp 108147 2025-01-31 14:05:57Z valery.portnyagin@oracle.com $ */
 /** @file
  * VBoxManage - Implementation of guestcontrol command.
  */
@@ -1565,7 +1565,7 @@ static RTEXITCODE gctlHandleRunCommon(PGCTLCMDCTX pCtx, int argc, char **argv, b
                         RTPrintf(GuestCtrl::tr("Exit code=%u (Status=%u [%s])\n"),
                                  lExitCode, procStatus, gctlProcessStatusToText(procStatus));
 
-                    rcExit = gctlRunCalculateExitCode(procStatus, lExitCode, true /*fReturnExitCodes*/);
+                    rcExit = gctlRunCalculateExitCode(procStatus, (ULONG)lExitCode, true /*fReturnExitCodes*/);
                 }
                 else if (   procStatus == ProcessStatus_TimedOutKilled
                          || procStatus == ProcessStatus_TimedOutAbnormally)
@@ -1703,7 +1703,7 @@ static RTEXITCODE gctlHandleCopy(PGCTLCMDCTX pCtx, int argc, char **argv, bool f
     }
 
     char **papszSources = RTGetOptNonOptionArrayPtr(&GetState);
-    size_t cSources = &argv[argc] - papszSources;
+    size_t cSources = (size_t)(&argv[argc] - papszSources);
 
     if (!cSources)
         return errorSyntax(GuestCtrl::tr("No sources specified!"));
